pyproject.toml
[tool.poetry]
name = "flask-base"
version = "0.1.0"
description = ""
authors = ["Takaaki Mizuno <takaaki.mizuno@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.1"
flask_script = "^2.0.6"
flask_injector = "^0.12.0"
flask_cors = "^3.0.8"
python-dotenv = "^0.10.3"
flask_sqlalchemy = "^2.4.1"
psycopg2 = "^2.8.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
isort = "^4.3.21"
yapf = "^0.29.0"
flake8 = "^3.7.9"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
